backend/requirements.txt
django
mysqlclient
pillow
django-ninja
django-cors-headers
requests
orjson

django-redis
django-celery-results
django-celery-beat
django-inlinecss

celery==5.2.3
redis
gunicorn

django-ninja-jwt
python-jose[cryptography]

python-dotenv